MAINT Split js2python_init into a separate .js file (#2136)
